+++ src/input/commodity/demand_slicing.rs
@@ -1,6 +1,7 @@
 //! Demand slicing determines how annual demand is distributed across the year.
 use super::super::*;
 use super::demand::*;
+use crate::id::IDCollection;
 use crate::time_slice::{TimeSliceID, TimeSliceInfo};
 use anyhow::{ensure, Context, Result};
 use itertools::Itertools;
